+++ frontend_v2/src/app/app-routing.module.ts
@@ -41,10 +41,6 @@ const routes: Routes = [
     loadChildren: './components/contact/contact.module#ContactModule',
   },
   {
-    path: 'analytics',
-    loadChildren: './components/analytics/analytics.module#AnalyticsModule',
-  },
-  {
     path: 'get-involved',
     loadChildren: './components/get-involved/get-involved.module#GetInvolvedModule',
   },
