--- frontend_v2/src/app/components/analytics/analytics.component.html
@@ -1,5 +1,0 @@
-<div class="web-container">
-  <div class="analytics-content">
-    <app-host-analytics></app-host-analytics>
-  </div>
-</div>
