--- frontend_v2/src/app/components/analytics/analytics.component.scss
@@ -1,25 +1,0 @@
-@import 'styles/variables';
-
-.analytics-content {
-  margin-left: 2%;
-}
-
-.web-container {
-  width: calc(100vw - 223px);
-  float: right;
-  overflow-x: hidden;
-  &.center {
-    float: none;
-    margin: 0 auto;
-    text-align: left;
-    overflow: hidden;
-  }
-}
-
-/*media queries*/
-
-@media only screen and (max-width: $med-screen) {
-  .web-container {
-    width: 100%;
-  }
-}
