+++ frontend_v2/src/app/components/challenge/challenge.module.ts
@@ -32,6 +32,7 @@ import { SharedModule } from '../../shared/shared.module';
 import { ChallengelistModule } from '../publiclists/challengelist/challengelist.module';
 import { SubmissionMetaAttributesDialogueComponent } from './submission-meta-attributes-dialogue/submission-meta-attributes-dialogue.component';
 import { ChallengediscussComponent } from './challengediscuss/challengediscuss.component';
+import { ChallengeanalyticsComponent } from './challengeanalytics/challengeanalytics.component';
 
 @NgModule({
   declarations: [
@@ -48,6 +49,7 @@ import { ChallengediscussComponent } from './challengediscuss/challengediscuss.c
     PhasecardComponent,
     SubmissionMetaAttributesDialogueComponent,
     ChallengediscussComponent,
+    ChallengeanalyticsComponent,
   ],
   imports: [
     CommonModule,
