+++ frontend_v2/src/app/components/challenge/challengeanalytics/challengeanalytics.component.spec.ts
@@ -1,0 +1,24 @@
+import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { ChallengeanalyticsComponent } from './challengeanalytics.component';
+
+describe('ChallengeanalyticsComponent', () => {
+  let component: ChallengeanalyticsComponent;
+  let fixture: ComponentFixture<ChallengeanalyticsComponent>;
+
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      declarations: [ChallengeanalyticsComponent],
+    }).compileComponents();
+  }));
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(ChallengeanalyticsComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});
