+++ frontend_v2/src/app/components/publiclists/challengelist/challengelist.component.html
@@ -85,10 +85,6 @@
           </div>
         </div>
       </mat-tab>
-      <!-- Analytics -->
-      <mat-tab label="Analytics">
-        <app-analytics></app-analytics>
-      </mat-tab>
     </mat-tab-group>
   </div>
   <!-- my participated challenges  -->
