+++ frontend_v2/src/styles/base.scss
@@ -1359,7 +1359,7 @@ code {
         color: $med-black;
         padding: 10px;
         font-size: $fs-18;
-        font-weight: $fw-bolder;
+        font-weight: 300;
       }
       .content {
         padding: 10px;
